requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image==0.16.2
tensorflow==2.2.0
keras==2.3.1
opencv-python
h5py==2.10.0
imgaug
pyinstaller
protobuf==3.20.*
IPython[all]